requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
plaid-python==16.0.0
pandas==2.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
